fix: rearrange types
