+++ frontend/src/components/Book/index.ts
@@ -1,3 +1,2 @@
 export { useBooks } from "./BookContext";
 export { BooksContextProvider } from "./BookContext";
-export * from "./Book.types";
